feat: Added new Dagger weapon and changed Sword weapon
Next I need to add a way to pick a weapon, for example from the floor and with text pop up

+++ Assets/Code/Fighter.cs
@@ -5,8 +5,8 @@ using UnityEngine;
 public abstract class Fighter : MonoBehaviour
 {  
     // Public fields
-    public int hitpoint = 10;
-    public int maxHitpoint = 10;
+    public float hitpoint = 10;
+    public float maxHitpoint = 10;
     public float pushRecoverySpeed = 1f;
 
     // Imunity
